hotfix: deleteCmd incorrectly syncing causing faulty content deletion
